CDHSH-77 Refactored the ProjectCard out into its own component.
