+++ services/frontend/src/components/ProjectCard/index.js
@@ -1,0 +1,1 @@
+export {default} from "./ProjectCard";
