+++ services/frontend/src/components/index.js
@@ -5,4 +5,5 @@ export {default as FormError} from "./FormError";
 export {default as LoadingButton} from "./LoadingButton";
 export {default as Navbar} from "./Navbar";
 export {default as NavSidebar} from "./NavSidebar";
+export {default as ProjectCard} from "./ProjectCard";
 export {default as ScrollContainer} from "./ScrollContainer";
