+++ services/frontend/src/scenes/Projects/Projects.scss
@@ -39,78 +39,7 @@
     flex-direction: column;
     padding: $default-spacing * 2;
 
-    .project-card {
+    .projects-list-card {
         margin: $default-spacing 0;
-    }
-}
-
-.project-card {
-    display: flex;
-    flex-shrink: 0;
-    max-width: $default-spacing * 34;
-
-    .project-card-content {
-        display: flex;
-        padding: $default-spacing*1.5 $default-spacing*2;
-
-        .project-card-activity-section {
-            align-self: flex-start;
-            display: flex;
-            flex: 1;
-        }
-
-        .project-card-content-section {
-            flex: 3;
-
-            .project-card-name {
-                font-size: $font-size-subtitle;
-                margin-bottom: $default-spacing;
-            }
-
-            .project-card-description {
-                color: $color-text-secondary;
-            }
-
-            .project-card-skills {
-                display: flex;
-                flex-wrap: wrap;
-                margin: 0 (-$default-spacing/2);  /* To account for the side padding of the skill-badges. */
-                margin-top: $default-spacing;     /* And yes, the brackets are needed. Otherwise, it renders as
-                                                     just "margin: -8px". */
-                .skill-badge {
-                    margin: $default-spacing / 2;
-                }
-
-                .project-card-more-skills {
-                    color: $color-text-secondary;
-                    font-weight: bold;
-                    margin: 0 $default-spacing/2;
-                    margin-top: $default-spacing;
-                }
-            }
-        }
-
-        .project-card-nav-section {
-            align-items: center;
-            display: flex;
-            margin-left: $default-spacing;
-            margin-right: -$default-spacing;
-
-            .project-card-nav {
-                color: $color-text-secondary;
-            }
-        }
-    }
-}
-
-.text-badge {
-    background-color: $color-neutral-8;
-    border-radius: $default-border-radius;
-    color: $color-text-primary;
-    padding: $default-spacing/2 $default-spacing;
-
-    &.text-badge--highlighted {
-        background-color: $color-secondary;
-        color: $color-light;
     }
 }
